refactor: extract large components out of page and modify routing
