+++ src/pages/settings/index.tsx
@@ -1,0 +1,3 @@
+export default function SettingsPage() {
+  return <h1>Placeholder</h1>;
+}
