+++ src/pages/settings/village-codes/types.ts
@@ -1,0 +1,16 @@
+import { VillageCode, NewVillageCode } from "@/db/schema";
+
+export const DEFAULT_FORM: NewVillageCode = {
+  code: "",
+  name: "",
+  colorHex: "#3b82f6", // Default blue
+  isVisible: true,
+};
+
+export type FormFields = {
+  id?: number;
+  name: string;
+  code: string;
+  colorHex: string;
+  isVisible: boolean | undefined;
+};
